PB AT

+++ avaliacoes/src/main/java/com/pbtp1/shared/messaging/EventoProduto.java
@@ -4,13 +4,14 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record EventoProduto(
         Long produtoId,
         String nome,
         String descricao,
-        Double preco,
+        BigDecimal preco,
         Long categoriaId,
         Integer estoque,
         String status,
